Rename bust_tokens to tokens and update grammar for char/cast support
Rename bust_tokens.{hpp,cpp} to tokens.{hpp,cpp} and update all includes.
Add char, i8, i32 types, char literals, escape sequences, hex escapes,
and cast_expr (as) to the grammar specification.
