+++ bust/inc/parser.hpp
@@ -16,8 +16,8 @@
 #include <sstream>
 
 #include <ast/nodes.hpp>
-#include <bust_tokens.hpp>
 #include <exceptions.hpp>
+#include <tokens.hpp>
 #include <unordered_map>
 
 //****************************************************************************
