+++ bust/src/lexer.cpp
@@ -19,9 +19,9 @@
 #include <string_view>
 #include <unordered_map>
 
-#include "bust_tokens.hpp"
 #include "exceptions.hpp"
 #include "token.hpp"
+#include "tokens.hpp"
 
 //****************************************************************************
 namespace bust {
