+++ bust/src/parser.cpp
@@ -21,11 +21,11 @@
 #include <vector>
 
 #include "ast/types.hpp"
-#include "bust_tokens.hpp"
 #include "lexer_interface.hpp"
 #include "operators.hpp"
 #include "source_location.hpp"
 #include "token.hpp"
+#include "tokens.hpp"
 #include "types.hpp"
 
 //****************************************************************************
